requirements.txt
anyio==3.6.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.1
beautifulsoup4==4.11.1
bidict==0.22.0
binaryornot==0.4.4
bleach==6.1.0
certifi==2022.5.18.1
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
comm==0.2.0
contourpy==1.2.0
cookiecutter==2.1.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
exceptiongroup==1.1.3
executing==2.0.1
fastjsonschema==2.18.1
filelock==3.7.1
Flask==2.1.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.1
flask-sock==0.5.2
Flask-SocketIO==5.2.0
Flask-SQLAlchemy==2.5.1
fonttools==4.44.0
fqdn==1.5.1
fuzzywuzzy==0.18.0
gevent==21.12.0
greenlet==1.1.2
h11==0.13.0
httptools==0.5.0
huggingface-hub==0.7.0
idna==3.3
importlib-metadata==4.11.4
importlib-resources==6.1.1
ipykernel==6.26.0
ipython==8.17.2
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jinja2-time==0.2.0
joblib==1.1.0
Js2Py==0.71
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.1
kiwisolver==1.4.5
MarkupSafe==2.1.1
matplotlib==3.8.1
matplotlib-inline==0.1.6
mistune==3.0.2
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.7
notebook_shim==0.2.3
numpy==1.22.4
overrides==7.4.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
Pillow==10.1.0
pipwin==0.5.2
platformdirs==4.0.0
prometheus-client==0.18.0
prompt-toolkit==3.0.40
psutil==5.9.6
pure-eval==0.2.2
PyAudio @ file:///C:/Users/ax_el/pipwin/PyAudio-0.2.11-cp39-cp39-win_amd64.whl
pycparser==2.21
pydantic==1.10.2
Pygments==2.16.1
pyjsparser==2.7.1
PyJWT==2.4.0
pyparsing==3.0.9
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.8.2
python-dotenv==0.21.0
python-engineio==4.3.2
python-json-logger==2.0.7
python-Levenshtein==0.12.2
python-slugify==6.1.2
python-socketio==5.6.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32==306
pywinpty==2.0.12
PyYAML==6.0
pyzmq==25.1.1
referencing==0.30.2
regex==2022.6.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.12.0
Send2Trash==1.8.2
simple-websocket==0.5.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
SQLAlchemy==1.4.37
SQLAlchemy-Utils==0.38.2
stack-data==0.6.3
starlette==0.20.4
terminado==0.18.0
text-unidecode==1.3
tinycss2==1.2.1
tokenizers==0.12.1
tomli==2.0.1
torch==1.11.0
tornado==6.3.3
tqdm==4.64.0
traitlets==5.13.0
transformers==4.19.4
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
uri-template==1.3.0
urllib3==1.26.9
watchfiles==0.17.0
wcwidth==0.2.9
webcolors==1.13
webencodings==0.5.1
websocket==0.2.1
websocket-client==1.6.4
websockets==10.3
Werkzeug==2.1.2
wsproto==1.1.0
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
